VRMShaders.TextureFactory
