+++ Assets/UniGLTF/Runtime/UniGLTF/IO/TextureIO/GltfTextureImporter.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 58277bcba49b13142acfa67d40041970
+guid: c9128dde9a44e1f43a808330ebfba145
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
