+++ Assets/VRM/Runtime/IO/MToonTextureParam.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: ab2b998b9235dc94a90ccaf2e40a50a6
+guid: a7af3a3bafe41a64db660a18b1c25097
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
